fix: extract table-spans.ts to break import cycle
chunk.ts -> tables.ts -> ir.ts -> chunk.ts formed a runtime import cycle.
Move parseTableSpans / isSafeTableBreak to a standalone table-spans.ts that
only depends on fences.ts, breaking the cycle.
